+++ test/datalevin/search_test.clj
@@ -244,15 +244,13 @@
 (deftest search-kv-test
   (let [dir    (u/tmp-dir (str "search-kv-" (UUID/randomUUID)))
         lmdb   (l/open-kv dir)
-        engine (sut/new-search-engine lmdb {:index-position? true})]
-    (l/open-dbi lmdb "raw")
-    (l/transact-kv
-      lmdb
-      [[:put "raw" 1 "The quick red fox jumped over the lazy red dogs."]
-       [:put "raw" 2 "Mary had a little lamb whose fleece was red as fire."]
-       [:put "raw" 3 "Moby Dick is a story of a whale and a man obsessed."]])
-    (doseq [i [1 2 3]]
-      (sut/add-doc engine i (l/get-value lmdb "raw" i)))
+        engine (sut/new-search-engine lmdb {:index-position? true
+                                            :include-text?   true})
+        texts  {1 "The quick red fox jumped over the lazy red dogs."
+                2 "Mary had a little lamb whose fleece was red as fire."
+                3 "Moby Dick is a story of a whale and a man obsessed."}]
+
+    (doseq [i [1 2 3]] (sut/add-doc engine i (texts i)))
 
     (is (not (sut/doc-indexed? engine 0)))
     (is (sut/doc-indexed? engine 1))
@@ -263,9 +261,19 @@
     (is (= (sut/search engine "red" ) [1 2]))
     (is (= (sut/search engine "red" {:display :offsets})
            [[1 [["red" [10 39]]]] [2 [["red" [40]]]]]))
+    (is (= (sut/search engine "red" {:display :texts})
+           [[1 "The quick red fox jumped over the lazy red dogs."]
+            [2 "Mary had a little lamb whose fleece was red as fire."]]))
+    (is (= (sut/search engine "red" {:display :texts+offsets})
+           [[1  "The quick red fox jumped over the lazy red dogs."
+             [["red" [10 39]]]]
+            [2  "Mary had a little lamb whose fleece was red as fire."
+             [["red" [40]]]]]))
+
     (testing "update"
       (sut/add-doc engine 1 "The quick fox jumped over the lazy dogs.")
       (is (= (sut/search engine "red" ) [2])))
+
     (testing "parallel update"
       (dorun (pmap #(sut/add-doc engine %1 %2)
                    [2 1 4]